Se agregan validadores

+++ src/main/java/com/tortu/api/rest/restservices/UsuarioRestService.java
@@ -3,7 +3,9 @@ package com.tortu.api.rest.restservices;
 import com.tortu.api.models.Usuario;
 import com.tortu.api.rest.mappers.UsuarioResourceMapper;
 import com.tortu.api.rest.resources.UsuarioResource;
+import com.tortu.api.rest.validators.GenericValidator;
 import com.tortu.api.services.UsuarioService;
+import com.tortu.api.utils.GeneralException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -23,10 +25,26 @@ import java.util.List;
 public class UsuarioRestService {
 
     @Autowired
+    @Qualifier("createUsuarioValidator")
+    private GenericValidator createUsuarioValidator;
+
+    @Autowired
+    @Qualifier("findUsuarioByIdValidator")
+    private GenericValidator findUsuarioByIdValidator;
+
+    @Autowired
+    @Qualifier("updateUsuarioValidator")
+    private GenericValidator updateUsuarioValidator;
+
+
+    @Autowired
     private UsuarioService usuarioService;
 
     @Autowired
     private UsuarioResourceMapper converter;
+
+    @Autowired
+    private UsuarioResourceMapper usuarioResourceMapper;
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
@@ -34,7 +52,7 @@ public class UsuarioRestService {
         List<Usuario>usuarioList = usuarioService.findAllUsuarios();
         List<UsuarioResource> usuarioResourceList = new ArrayList<UsuarioResource>();
         for(Usuario usuario:usuarioList){
-            UsuarioResource resource = new UsuarioResource(usuario);
+            UsuarioResource resource = usuarioResourceMapper.mapper(usuario);
             usuarioResourceList.add(resource);
         }
         return  Response.ok(usuarioResourceList).build();
@@ -44,6 +62,7 @@ public class UsuarioRestService {
     @Path("/{usuario}")
     @Produces({MediaType.APPLICATION_JSON})
     public Response findUsuarioById(@PathParam("usuario") Usuario usuario) {
+        findUsuarioByIdValidator.validate(usuario);
         Usuario usuarioResponse =  usuarioService.findUsuario(usuario);
         UsuarioResource resource = converter.mapper(usuario);
         return Response.ok(resource).build();
@@ -52,8 +71,9 @@ public class UsuarioRestService {
     @POST
     @Produces(MediaType.APPLICATION_JSON)
     public Response createUsuario(Usuario usuario){
+        createUsuarioValidator.validate(usuario);
         Usuario usuarioResponse = usuarioService.saveUsuario(usuario);
-        UsuarioResource resource = new UsuarioResource(usuarioResponse);
+        UsuarioResource resource = usuarioResourceMapper.mapper(usuarioResponse);
         return Response.ok(resource).build();
     }
 
@@ -61,14 +81,18 @@ public class UsuarioRestService {
     @Path("/usuario")
     @Produces(MediaType.APPLICATION_JSON)
     public Response updateUsuario( @PathParam("usuario") Usuario usuario ){
+        updateUsuarioValidator.validate(usuario);
         Usuario usuarioResponse = usuarioService.updateUsuario(usuario);
-        UsuarioResource resource = new UsuarioResource(usuarioResponse);
+        UsuarioResource resource = usuarioResourceMapper.mapper(usuarioResponse);
         return Response.ok(resource).build();
     }
 
     @DELETE
     @Path("/usuarioId")
     public Response deleteUsuario(@PathParam("usuarioId") Integer id){
+        if(id==null){
+            throw new GeneralException("El ID del usuario es nulo");
+        }
         usuarioService.deleteUsuario(id);
         return Response.ok().build();
     }
